first commit to refactore dev branch
