+++ README.md
@@ -1,3 +1,7 @@
 # Portfolio Website
 
 You can visit my portfolio website [here](https://portfolio-lilac-psi.vercel.app/)
+
+## Refactore List
+- [ ] Start by using different branches for production and development
+- [ ] Move the project to typescript.
