+++ src/components/TypeEffect.js
