+++ src/components/navbar.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Button, 
 Box, 
-makeStyles,
 Typography,
 AppBar,
 Toolbar,
@@ -13,42 +12,8 @@ Divider,
 Link
 } from '@material-ui/core';
 import MenuIcon from '@material-ui/icons/Menu';
-
+import useStyles from './test';
 const NavBar = () => {
-
-    const useStyles = makeStyles((theme) => ({
-        hideButtons: {
-            [theme.breakpoints.down('xs')]:{
-                display:"none"
-            }
-          },
-          hideIconMenu: {
-            [theme.breakpoints.up('sm')]:{
-                display: "none"
-            }
-          },
-        button:{
-            background: "inherit",
-            fontSize: "0.9rem",
-            height: 40,
-            padding: "0 1rem",
-            color: "black",
-        },
-        title: {
-          color: "Black"
-        },
-        link: {
-            textDecoration: 'none',
-            flex: 1,
-            '&:hover':{
-              cursor: 'pointer',
-              textDecoration: "none",
-            }
-        },
-        appbar: {
-            background: "#ffffff"
-        }
-    }));
     
     const scrollTo = (id) => {
         const element = document.getElementById(id);
